% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter of a string}
\usage{
capitalize(string)
}
\arguments{
\item{string}{Any string}
}
\value{
Input string with the first letter of the first word capitalized
}
\description{
Credit to the Hmisc package
}
\examples{
capitalize("scott")
capitalize("hello there")

}
