% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize_all}
\alias{capitalize_all}
\title{Capitalize the first letter of each word in a string}
\usage{
capitalize_all(string)
}
\arguments{
\item{string}{Any string}
}
\value{
Input string with the first letter of each word capitalized
}
\description{
Credit to the Hmisc package
}
\examples{
capitalize("scott")
capitalize("hello there")

}
