% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{randomizer}
\alias{randomizer}
\title{Create a randomizer column on a data frame}
\usage{
randomizer(data, smp_prop)
}
\arguments{
\item{data}{A data frame}

\item{smp_prop}{Sample size proportion to be selected}
}
\description{
TRUE/FALSE indicator on a dataset that shows if that row is to be selected
or not. Can be used in train/test validation.
}
