% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMarkdown.R
\name{runAllChunks}
\alias{runAllChunks}
\title{Run All Chunks}
\usage{
runAllChunks(rmd, envir = globalenv())
}
\description{
Run all chuncks of code in a R markdown file and load them into the
console for exploration and development
}
