% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{Take a dataset and split it randomly into train and test sets
The size of the training set is the smp_size argument}
\usage{
splitData(data, smp_size)
}
\arguments{
\item{smp_size}{A proportion or number of rows to be in the training set}
}
\value{
A named list of datasets with the original train and test
}
\description{
Take a dataset and split it randomly into train and test sets
The size of the training set is the smp_size argument
}
