% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpose.R
\name{tpose}
\alias{tpose}
\title{Transpose a data.frame preserving names}
\usage{
tpose(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Takes in a data.frame and returns it transposed, with the original row names as column names and the original column names as row names.
}
\examples{
tpose(mtcars[1:5, 1:5])
}
